move projects_v2 into the designsafe api module
