+++ designsafe/apps/api/projects_v2/schema_models/base.py
@@ -7,13 +7,13 @@ from pydantic import (
     model_validator,
     Field,
 )
-from designsafe.apps.projects_v2.constants import (
+from designsafe.apps.api.projects_v2.constants import (
     NATURAL_HAZARD_TYPES,
     FIELD_RESEARCH_TYPES,
     OTHER_DATA_TYPES,
 )
-from designsafe.apps.projects_v2.schema_models._field_models import MetadataModel
-from designsafe.apps.projects_v2.schema_models._field_models import (
+from designsafe.apps.api.projects_v2.schema_models._field_models import MetadataModel
+from designsafe.apps.api.projects_v2.schema_models._field_models import (
     AssociatedProject,
     DropdownValue,
     FileTag,
@@ -24,7 +24,7 @@ from designsafe.apps.projects_v2.schema_models._field_models import (
     ProjectUser,
     ReferencedWork,
 )
-from designsafe.apps.projects_v2.schema_models._field_transforms import (
+from designsafe.apps.api.projects_v2.schema_models._field_transforms import (
     handle_array_of_none,
     handle_award_number,
     handle_dropdown_value,
